backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
bcrypt
google-generativeai
SQLAlchemy==2.0.36
flask
flask_cors
elevenlabs
opencv-python>=4.5.5
mediapipe>=0.8.9
numpy>=1.21